# openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/habitTrackerAPI/user/updateUser/{id}":{"put":{"tags":["user-controller"],"summary":"Useri update edir","description":"Verilen id-e gore useri tapib yenileyir","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/habitTrackerAPI/habit/updateHabit/{id}":{"put":{"tags":["habit-controller"],"summary":"Habiti yenilemek","description":"Verilen id-ye gore habiti yenileyir","operationId":"updateHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HabitsResponse"}}}}}}},"/habitTrackerAPI/user/createUser":{"post":{"tags":["user-controller"],"summary":"Yeni User yaratmaq","description":"Yeni User yaradir","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"201":{"description":"Created","headers":{"Location":{"schema":{"type":"string"},"description":"URI of the newly created resource"}},"content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/habitTrackerAPI/user/calculate-level/{userId}":{"post":{"tags":["user-controller"],"summary":"Habit tamamlanma Leveli","description":"Habiti nece tamamlandiginizi yoxlayir","operationId":"calculateHabitLevel","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/habitTrackerAPI/habitEntry/markHabitAsCompleted":{"post":{"tags":["habit-entry-controller"],"summary":"Habiti tamamlayir","description":"Habitin tamamlandigini gosterir (Streak-i yeniləyir)","operationId":"markHabitAsCompleted","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitEntryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HabitEntryResponse"}}}}}}},"/habitTrackerAPI/habit/createHabit":{"post":{"tags":["habit-controller"],"summary":"Yeni Habit yaratmaq","description":"Yeni Habit yaradir","operationId":"createHabit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HabitsResponse"}}}}}}},"/habitTrackerAPI/user/getUserById/{id}":{"get":{"tags":["user-controller"],"summary":"Useri tapir","description":"Verilen id-e gore useri tapir","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/habitTrackerAPI/user/getAllUsers":{"get":{"tags":["user-controller"],"summary":"Bütün Userləri verir","description":"Bütün Userləri səhifələnmiş formada göstərir","operationId":"getAllUsers","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserResponse"}}}}}}},"/habitTrackerAPI/habitEntry/getEntriesByHabitId/{id}":{"get":{"tags":["habit-entry-controller"],"summary":"Entryleri getirir","description":"Habit id sine gore entryleri getirir","operationId":"getEntriesByHabitId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HabitEntryResponse"}}}}}}}},"/habitTrackerAPI/habit/getHabitById/{id}":{"get":{"tags":["habit-controller"],"summary":"Habiti getirir","description":"Verilen id ye gore habiti getirir","operationId":"getHabitById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Habits"}}}}}}},"/habitTrackerAPI/habit/getAllHabits":{"get":{"tags":["habit-controller"],"summary":"Butun habitleri getirir","description":"Butun habitleri getirir","operationId":"getAllHabits","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HabitsResponse"}}}}}}}},"/habitTrackerAPI/user/deleteUser/{id}":{"delete":{"tags":["user-controller"],"summary":"User silir","description":"Verilen id-e gore User-i silir","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/habitTrackerAPI/habitEntry/deleteEntry/{id}":{"delete":{"tags":["habit-entry-controller"],"summary":"Entryni silir","description":"Entryni silir","operationId":"deleteEntry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/habitTrackerAPI/habit/deleteHabit/{id}":{"delete":{"tags":["habit-controller"],"summary":"Habit silmek","description":"Verilen id-ye gore habiti silir","operationId":"deleteHabit","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Habits":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"frequency":{"type":"string","enum":["DAILY","WEEKLY","MONTHLY","CUSTOM"]},"habitName":{"type":"string"},"habitDescription":{"type":"string"},"currentStreak":{"type":"integer","format":"int32"},"longestStreak":{"type":"integer","format":"int32"},"user":{"$ref":"#/components/schemas/User"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"habits":{"type":"array","items":{"$ref":"#/components/schemas/Habits"}}}},"UserRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":30,"minLength":1},"email":{"type":"string"},"habits":{"type":"array","items":{"$ref":"#/components/schemas/Habits"},"maxItems":30,"minItems":1}},"required":["habits","name"]},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"habits":{"type":"array","items":{"$ref":"#/components/schemas/Habits"}}}},"HabitsRequest":{"type":"object","properties":{"frequency":{"type":"string","enum":["DAILY","WEEKLY","MONTHLY","CUSTOM"]},"habitName":{"type":"string","maxLength":30,"minLength":1},"habitDescription":{"type":"string","maxLength":30,"minLength":1},"user":{"$ref":"#/components/schemas/User"}},"required":["frequency","habitDescription","habitName","user"]},"HabitsResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"habitName":{"type":"string"},"habitDescription":{"type":"string"},"user":{"$ref":"#/components/schemas/User"},"frequency":{"type":"string","enum":["DAILY","WEEKLY","MONTHLY","CUSTOM"]},"currentStreak":{"type":"integer","format":"int32"},"longestStreak":{"type":"integer","format":"int32"}}},"HabitEntryRequest":{"type":"object","properties":{"habitId":{"$ref":"#/components/schemas/Habits"},"date":{"type":"string","format":"date"},"userId":{"type":"integer","format":"int64"}},"required":["date","habitId","userId"]},"HabitEntryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"habitId":{"$ref":"#/components/schemas/Habits"},"date":{"type":"string","format":"date"},"completed":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageUserResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/Sortnull"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/Pageablenull"},"empty":{"type":"boolean"}}},"Pageablenull":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/Sortnull"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"Sortnull":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}}}}